Add support for igored fields
Right now all fiedlds are used for generation as long as they are exported.

There are many occations where the struct can contains fields which are not mapped to database  columns for instance

```go
type Foo struct {
	ID   int64
	Data map[string]interface{} `sql:"-"`
}
```

The Data field is not supposed to be mapped to  any database columns.

So, adding a `sql:"-"` tag will skip/ignore  the field.